+++ packages/expo-local-authentication/android/src/main/java/expo/modules/localauthentication/LocalAuthenticationModule.kt
@@ -206,6 +206,8 @@ class LocalAuthenticationModule : Module() {
     }
 
     val promptMessage = options.promptMessage
+    val promptSubtitle = options.promptSubtitle
+    val promptDescription = options.promptDescription
     val cancelLabel = options.cancelLabel
     val requireConfirmation = options.requireConfirmation
     val allowedAuthenticators = if (options.disableDeviceFallback) {
@@ -220,6 +222,8 @@ class LocalAuthenticationModule : Module() {
     biometricPrompt = BiometricPrompt(fragmentActivity, executor, authenticationCallback)
     val promptInfoBuilder = PromptInfo.Builder().apply {
       setTitle(promptMessage)
+      setSubtitle(promptSubtitle)
+      setDescription(promptDescription)
       setAllowedAuthenticators(allowedAuthenticators)
       if (options.disableDeviceFallback) {
         setNegativeButtonText(cancelLabel)
@@ -248,6 +252,8 @@ class LocalAuthenticationModule : Module() {
     }
 
     val promptMessage = options.promptMessage
+    val promptSubtitle = options.promptSubtitle
+    val promptDescription = options.promptDescription
     val requireConfirmation = options.requireConfirmation
 
     // BiometricPrompt callbacks are invoked on the main thread so also run this there to avoid
@@ -255,7 +261,7 @@ class LocalAuthenticationModule : Module() {
     appContext.mainQueue.launch {
       // On Android devices older than 11, we need to use Keyguard to unlock by Device Credentials.
       if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
-        val credentialConfirmationIntent = keyguardManager.createConfirmDeviceCredentialIntent(promptMessage, "")
+        val credentialConfirmationIntent = keyguardManager.createConfirmDeviceCredentialIntent(promptMessage, promptDescription)
         fragmentActivity.startActivityForResult(credentialConfirmationIntent, DEVICE_CREDENTIAL_FALLBACK_CODE)
         return@launch
       }
@@ -267,6 +273,8 @@ class LocalAuthenticationModule : Module() {
 
       val promptInfoBuilder = PromptInfo.Builder().apply {
         setTitle(promptMessage)
+        setSubtitle(promptSubtitle)
+        setDescription(promptDescription)
         setAllowedAuthenticators(BiometricManager.Authenticators.DEVICE_CREDENTIAL)
         setConfirmationRequired(requireConfirmation)
       }